Bug 14697: Extend and enhance "Claims returned" lost status
This adds a "Claims returned" feature that extends and enhances the claims returned lost status

Test Plan:
1) Create a "Claims Returned" lost value
2) Create some RETURN_CLAIM_RESOLUTION authorized values
3) Set ClaimReturnedLostValue
4) Set ClaimReturnedChargeFee
5) Set ClaimReturnedWarningThreshold
6) Create some checkouts
7) Claim some returns
8) Verify ClaimReturnedChargeFee works with all 3 options
9) Verify ClaimReturnedWarningThreshold shows a warning once the threshold has been exceeded
10) Edit notes on a claim
11) Resolve a claim
12) Delete a claim

Signed-off-by: Andrew Fuerste-Henry <andrew@bywatersolutions.com>

Signed-off-by: Lisette Scheer <lisetteslatah@gmail.com>

Bug 14697: Show an item has unresolved claims on checkin

Signed-off-by: Andrew Fuerste-Henry <andrew@bywatersolutions.com>

Signed-off-by: Lisette Scheer <lisetteslatah@gmail.com>

Bug 14697: Move /return_claims/claim to /return_claims

Bug 14697: Don't unbless claims before rendering, it's done automatically

Bug 14697: Use try/catch blocks

Bug 14697: Add exceptions and overload ReturnClaim->store

This patch adds an overloaded Koha::Checkouts::ReturnClaim->store method
that will raise an exception if the mandatory created_by attribute is
missing.

Tests are added for this overloaded method.

To test:
1. Apply this patch
2. Run:
   $ kshell
  k$ prove t/db_dependent/Koha/Checkouts/ReturnClaim.t
=> SUCCESS: Tests pass
3. Sign off :-D

Bug 14697: Wrap claim_returned logic inside a transaction

This patch puts all the logic inside a transaction and also an outer
try/catch construct.

Koha::Exceptions::Exception-derived exceptions are rethrown.

TODO: Maybe we want to handle specific cases in a custom handy way.
DBIx::Class exceptions will probably be wrapped in
Koha::Exception::Object-based exceptions (notably FK failures, etc).
TODO: Think of the exceptions in terms of how they will be used to
return response codes and messages on the API

Bug 14697: Some API fixes

This fixes focus on:
- Attribute names consistency
- Return code for POST (201)
- Some special handling for dates/date-times on the spec

TODO: Should created_by be a parameter or should it be 'detected' from
the 'session' or 'authentication mechanism'?
TODO: Some checks and design smell a bit. If the DB can have a UNIQUE
constraint on issue_id, why not just create the claim here and let the
ORM throw an exception if the checkout doesn't exist or is duplicated?
Non-existent patron or librarian will be the same as they are FK.

Instead of having Koha::Checkout->claim_returned we could do it all in
the controller. This is related to the above thing. I'm not against the
current approach, but got me thinking.

There is usually a common doubt when designing the API. Conceptually,
one would say POST /return_claims { object } is the way to do it.

But the class design suggests it could also be POST
/checkouts/{checkout_id}/return_claim { object }, even if /return_claims
exists too (maybe for listing them). If the entry point is by
checkout_id, the Location header should be set
/api/v1/return_claims/{claim_id} anyway, so other actions against it
have the correct path (i.e. if we need a PUT or whatever for the other
actions). I will add another patch implementing this.

I don't have the answer, even if it wasn't this late. But it is
something to think about.

Conclusion: bailout the lowest possible level, throwing an exception,
rely on exceptions through try/catch constructs. Use of transactions.
On the API it is prefered to just return a 500 with the stringified exception
It is easier to code. The best thing to do is to know in advance the
possible exceptions, and handle them accordingly (for example, if
there's a FK violation for the 'parent' checkout, return a 404, if it is
violating a UNIQUE, return 400 (bad request). And so on.

Bug 14697: 'Guess' the staff user

Instead of using C4::Context->userenv (which is cookie-only) in the
Koha::Object layer, we just do it in the controllers. Easier to test,
cleaner in terms of design (why would this business-ish level that deals
with checkout DB rows need to know about sessions? cookie?).

So in the controllers we do have that info, and the API does it pretty
good: $c->stash('koha.user') will give you a Koha::Patron object matching
the user that is interacting with the API, no matter what auth method is
using.

Keep in mind that this Perl API might be used in other places that don't
have userenv (cronjobs?).

Bug 14697: Add Location header on POST response

The changes on the API are not reflected on the tests! Let me know if
this is something you need help with.

The 'complete' example is the libraries endpoint and its tests.

Remove old_issue_id column from return_claims table, change issue_id to UNIQUE constrain with no FK

Fix existing unit tests

CLEANUP

FIXES FOR STAFF INTERFACE