Bug 23165: Renewals for Self Checkout (SCO) module do not record branchcode in statistics
Koha considers SCO to be part of the OPAC, but the code related to OpacRenewalBranch is in opac-renew.pl. Because of this, the branchcode does not get set when a renewal takes place via the self check module. It would make more sense to move this logic into C4::Circulation::AddRenewal to centralize it and ensure that the statistics branch is set no matter where in the OPAC AddRenewal may be called.

Test Plan:
1) Using the SCO module, renew an item
2) View the statistics table row for the renewal, note that branch is empty
3) Apply this patch
4) Restart all the things!
5) Renew the item again
6) View the stats table row for the new renewal, the branch code should be set
   as per the system preference OpacRenewalBranch