Bug 26282: Allow staff to decide if a hold cancellation notice will be sent when cancelling a hold
Bug 25534 adds the option to send hold cancellation notices when a
reason was given. It would be nice if you could give a reason but still
decide not to send to the patron.

Say the patron called and will not be able to pick up numerous holds for
some time. You cancel them all with "patron requested cancellation" but
don't want to send multiple notices.

I could imagine a checkbox next to the reason to activate/deactivate
sending the notice.

Test Plan:
1) Apply this patch
2) Visit each area in Koha where a hold can be canceled with a reason
3) Note the new 'Notify patron' checkbox
4) Test canceling a hold with and without this checkbox checked
5) Verify leaving it unchecked does not trigger a notice to be sent to the patron

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>
Signed-off-by: Josef Moravec <josef.moravec@gmail.com>
Rebased-by: Victor Grousset/tuxayo <victor@tuxayo.net>