Bug 26692: Add barcode image generator service for OPAC
Some libraries would like to add a scannable barcode to the patron self
service area of the opac for patrons who do not have their library cards
with them. The existing svc/barcode on the staff side should work fine,
but requires a staff login. It makes sense to duplicate this script
but only require a valid patron login to use it.

Test Plan:
1) Apply this patch
2) Browse to /cgi-bin/koha/svc/barcode?barcode=123456789 from your OPAC address
3) Note the barcode image