Bug 26351: Add call_recursive() as a replacement for call()
The method call() is not sufficient for barcode transformations. It's
possible that more than one barcode transformation plugin will be
installed. call_recursive() takes the output of the first plugin and
uses it as the input for the next plugin and so on. This allowes each
plugin to see the current version of the barcode and modify it if
necessary.