Bug 24857: Adapt code to the volumes => item_groups change
This patch updates the code so all references to 'volumes' are
translated into 'item groups'.

Relevant changes:
- API routes and specs
- Permission name is changed
- Syspref name changed

Note: I removed a behavior that felt unnatural (at least in the place it
was put): deleting the item group if the linked item is deleted. It
feels like a controller thing, and also something that could be offered
to the staff on the UI on item deletion. To avoid blocking the feature,
I suggest this removal, and discuss it in its own context.

To test:
1. Run:
   $ reset_all
2. Apply this patchset
3. Run:
   $ updatedatabase
=> SUCCESS: Update works
4. Repeat 1 (with the patches applied)
=> SUCCESS: Install works with this DB structure
5. Run all the tests:
   $ kshell
  k$ git diff origin/master --name-only | grep -e '\.t$' | xargs prove
=> SUCCESS: Tests pass!
6. Follow the original plan (with the Volume=>ItemGRoup change in mind)
=> SUCCESS: All good!
7. Sign off :-D

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>