WIP - Bug 30529: Allow all items fields to be set in MARCItemFieldsToOrder
The MARCItemFieldsToOrder syspref allows values for a set list of item/order fields to be sent in MARC file used in acquisitions.
We should broaden that out to include any item field to support a broader array of workflows.

This patch allows any item column to be used in MARCCItemFieldsToOrder.
It continues to origonal misnamed fields 'loc' ( location ) and 'copyno' ( copynumber )

Test Plan:
1) Set up your MARCItemFieldsToOrder, verify everything is working
2) Add a new mapping to MarcItemFieldsToOrder for a previously
   unsupported field ( such as "materials" )